requirements.txt
discord.py
Flask
openai